manually set weights for VGG16 model base to ensure they will be present in production
These are the same weights from imagenet. The difference is that instead
of relying on Keras to make a network request and save them "somewhere",
we can specify the file containing them.

Keras by default saves the weights it downloads to ~/.keras/models/ , so
I copied the file from there.