added classification pipeline to tl_classifier.py
Right now it just logs the prediction, rather than returning it. This is for
debugging, and we need to get it actually working eventaully.

You also have to manually ensure it always runs in tl_detector.py, since that
code is incomplete and lacks the proper conditional to run the classifier code
just yet.

NOTE: I intentionally did not commit the weights for the VGG16 model, since
the weights file is very large. If you need it, you can run the classifier
and then copy the weights file from ~/.keras/models